feat: finaliza CRUD de vagas, implementa DTOs e tratamento global de erros (400 e 404)

+++ src/main/java/com/portal/emprego/service/VagaService.java
@@ -1,5 +1,7 @@
 package com.portal.emprego.service;
 
+import com.portal.emprego.dto.VagaRequestDTO;
+import com.portal.emprego.exception.RecursoNaoEncontradoException;
 import com.portal.emprego.model.Vaga;
 import com.portal.emprego.repository.VagaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,11 +24,27 @@ public class VagaService {
     }
 
     public Vaga salvar(Vaga vaga) {
-        // Aqui você pode colocar validações no futuro (ex: validar se a empresa existe)
+        // colocar validações no futuro (ex: validar se a empresa existe)
         return vagaRepository.save(vaga);
     }
+    // UPDATE
+    public Vaga atualizar(Long id, VagaRequestDTO dto) {
+        Vaga vagaExistente = vagaRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Vaga não encontrada com o ID: " + id));
 
+        vagaExistente.setTitulo(dto.getTitulo());
+        vagaExistente.setDescricao(dto.getDescricao());
+        vagaExistente.setEmpresa(dto.getEmpresa());
+        vagaExistente.setSalario(dto.getSalario());
+        vagaExistente.setLocalizacao(dto.getLocalizacao());
+
+        return vagaRepository.save(vagaExistente);
+    }
+    // DELETE: Remove a vaga se ela existir
     public void deletar(Long id) {
+        if (!vagaRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Não foi possível deletar. Vaga não encontrada com o ID: " + id);
+        }
         vagaRepository.deleteById(id);
     }
 }
